+++ src/main/java/org/seedstack/oauth/AccessTokenValidator.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.oauth.spi;
+package org.seedstack.oauth;
 
 /**
  * Implement this interface to provide a custom validator for OAuth access token. This is only required when NOT using