+++ src/main/java/org/seedstack/oauth/OAuthAuthenticationToken.java
@@ -5,13 +5,21 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.oauth.spi;
+package org.seedstack.oauth;
 
-import java.util.Optional;
 import org.seedstack.seed.security.AuthenticationToken;
 
+import java.util.Optional;
+
 public interface OAuthAuthenticationToken extends AuthenticationToken {
+    /**
+     * @return the access token as a string.
+     */
     String getAccessToken();
 
+    /**
+     * @return the refresh token as a string if any.
+     */
     Optional<String> getRefreshToken();
+
 }