+++ src/main/java/org/seedstack/oauth/OidcAuthenticationToken.java
@@ -5,8 +5,19 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.oauth.spi;
+package org.seedstack.oauth;
+
+import java.util.Map;
 
 public interface OidcAuthenticationToken extends OAuthAuthenticationToken {
+    /**
+     * @return the ID JSON Web Token (JWT) in its compact format consisting
+     * of Base64URL-encoded parts delimited by period ('.') characters.
+     */
+    String getIdToken();
 
+    /**
+     * @return all the claims present in the identity token.
+     */
+    Map<String, Object> getIdClaims();
 }