+++ src/main/java/org/seedstack/oauth/internal/OAuthAuthenticationTokenImpl.java
@@ -11,14 +11,20 @@ import com.google.common.base.Preconditions;
 import com.nimbusds.oauth2.sdk.token.AccessToken;
 import com.nimbusds.oauth2.sdk.token.RefreshToken;
 import org.apache.shiro.authc.AuthenticationToken;
-import org.seedstack.oauth.spi.OAuthAuthenticationToken;
+import org.seedstack.oauth.OAuthAuthenticationToken;
 
 import java.util.Optional;
 
 public class OAuthAuthenticationTokenImpl implements AuthenticationToken, OAuthAuthenticationToken {
+    static final OAuthAuthenticationTokenImpl EMPTY = new OAuthAuthenticationTokenImpl();
     private static final long serialVersionUID = 1L;
     private final AccessToken accessToken;
     private final RefreshToken refreshToken;
+
+    private OAuthAuthenticationTokenImpl() {
+        this.accessToken = null;
+        this.refreshToken = null;
+    }
 
     OAuthAuthenticationTokenImpl(AccessToken accessToken, RefreshToken refreshToken) {
         this.accessToken = Preconditions.checkNotNull(accessToken, "Access token cannot be null");
@@ -31,12 +37,12 @@ public class OAuthAuthenticationTokenImpl implements AuthenticationToken, OAuthA
     }
 
     @Override
-    public Object getPrincipal() {
+    public String getPrincipal() {
         return "";
     }
 
     @Override
-    public Object getCredentials() {
+    public AccessToken getCredentials() {
         return accessToken;
     }
 
@@ -50,14 +56,8 @@ public class OAuthAuthenticationTokenImpl implements AuthenticationToken, OAuthA
         return Optional.ofNullable(refreshToken).map(RefreshToken::getValue);
     }
 
-    /**
-     * Returns a textual representation of the authentication token. The access token is omitted for security
-     * purposes. The identity is unknown.
-     *
-     * @return the textual representation as {@link String}.
-     */
     @Override
     public String toString() {
-        return "OAuthAuthenticationToken[" + getPrincipal() + "]";
+        return "OAuthAuthenticationToken[?]";
     }
 }