+++ src/main/java/org/seedstack/oauth/internal/OAuthModule.java
@@ -9,9 +9,9 @@ package org.seedstack.oauth.internal;
 
 import com.google.inject.PrivateModule;
 import com.google.inject.util.Providers;
-import org.seedstack.oauth.spi.AccessTokenValidator;
-import org.seedstack.oauth.spi.OAuthProvider;
-import org.seedstack.oauth.spi.OAuthService;
+import org.seedstack.oauth.AccessTokenValidator;
+import org.seedstack.oauth.OAuthProvider;
+import org.seedstack.oauth.OAuthService;
 
 class OAuthModule extends PrivateModule {
     private final OidcDiscoveryDocument discoveryDocument;