+++ CHANGELOG.md
@@ -1,10 +1,17 @@
+# Version 3.0.0 (2020-11-23)
+
+* [brk] Moved everything under the `org.seedstack.oauth.spi` package to `org.seedstack.oauth`.
+* [chg] The application id is now used as the default allowed audience if none is configured.
+* [chg] Raw user info is available as implementation-specific subject principal of type `com.nimbusds.openid.connect.sdk.claims.UserInfo`.
+* [chg] Raw tokens are available as `org.seedstack.oauth.OAuthAuthenticationToken` subject principal.
+
 # Version 2.0.0 (2020-11-09)
 
 * [brk] Simplified and new configuration options.
 * [new] A default opaque token validator is now provided (validating the access token by calling the userInfo endpoint if available). 
 * [new] Improved token validation (at_hash optional, try validating access token as JWT first, then as an opaque token if it fails).
 * [new] Allow treating scopes either as direct permissions (the default) or as roles (which then can give permissions through a `RolePermissionResolver`).
-* [new] Extract claims from ID token if available and enrich them with the userInfo endpoint if enabled (off by default).
+* [new] Claims are now extracted from identity token and optionally enriched with the userInfo endpoint if enabled (off by default).
 * [new] Proper support for bearer access tokens.
 * [chg] Improved client error messages.